Have UDS listener use native AOT and trimming to reduce precompiled binary size from 91MB to 10MB
