+++ FirecrackerSharp.Host.Ssh.Tests/uds-listener.bin
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:03c852dec78af3e29bf93326353e0d88794e2a0920fbbc9fd8e80a3965aa51a0
-size 96049375
+oid sha256:283da48f7146997c024a36722ff8e29578d6eb08e1ce4bfe89a6ab1c03ffa078
+size 10983752
